backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
SQLAlchemy
Werkzeug
Authlib
deep-translator
psycopg2-binary
python-dotenv